transformers/generic-py/requirements.txt
torchvision==0.8.1
requests==2.22.0
torch==1.7.0
Pillow==8.0.1
